refactor: Modularizando controller
