+++ src/main/resources/br/ufrn/imd/circusmanager/tela1.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.control.Button?>
 
 <VBox alignment="CENTER" spacing="20.0" xmlns:fx="http://javafx.com/fxml"
-      fx:controller="br.ufrn.imd.circusmanager.Control.TelaInicialController">
+      fx:controller="br.ufrn.imd.circusmanager.Controller.CircoTelas.TelaInicialController">
     <padding>
         <Insets bottom="20.0" left="20.0" right="20.0" top="20.0"/>
     </padding>
