+++ src/main/resources/br/ufrn/imd/circusmanager/tela10.fxml
@@ -9,7 +9,7 @@
 <?import javafx.scene.layout.HBox?>
 
 <VBox alignment="CENTER" spacing="15.0" xmlns:fx="http://javafx.com/fxml"
-      fx:controller="br.ufrn.imd.circusmanager.Control.TelaAddAnimalController">
+      fx:controller="br.ufrn.imd.circusmanager.Controller.ZooTelas.TelaAddAnimalController">
     <padding>
         <Insets bottom="20.0" left="20.0" right="20.0" top="20.0"/>
     </padding>
