+++ src/main/resources/br/ufrn/imd/circusmanager/tela2.fxml
@@ -8,7 +8,7 @@
 <?import javafx.scene.control.TextField?>
 
 <StackPane xmlns:fx="http://javafx.com/fxml" 
-        fx:controller="br.ufrn.imd.circusmanager.Control.TelaLoginController" 
+        fx:controller="br.ufrn.imd.circusmanager.Controller.CircoTelas.TelaLoginController"
         prefWidth="800" prefHeight="600">
     <children>
         <VBox alignment="CENTER" spacing="15">
