+++ src/main/resources/br/ufrn/imd/circusmanager/tela3.fxml
@@ -3,7 +3,7 @@
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
 
-<StackPane xmlns:fx="http://javafx.com/fxml" fx:controller="br.ufrn.imd.circusmanager.Control.TelaCriarCircoController">
+<StackPane xmlns:fx="http://javafx.com/fxml" fx:controller="br.ufrn.imd.circusmanager.Controller.CircoTelas.TelaCriarCircoController">
     <children>
         <VBox alignment="CENTER" spacing="20">
             <!-- Título -->
