+++ src/main/resources/br/ufrn/imd/circusmanager/tela4.fxml
@@ -4,7 +4,7 @@
 <?import javafx.scene.layout.*?>
 
 <VBox alignment="CENTER" spacing="20.0" xmlns:fx="http://javafx.com/fxml"
-      fx:controller="br.ufrn.imd.circusmanager.Control.TelaCircoMenuController">
+      fx:controller="br.ufrn.imd.circusmanager.Controller.CircoTelas.TelaCircoMenuController">
     <padding>
         <Insets bottom="20.0" left="20.0" right="20.0" top="20.0"/>
     </padding>
