+++ src/main/resources/br/ufrn/imd/circusmanager/tela9.fxml
@@ -7,7 +7,7 @@
 <?import javafx.scene.layout.HBox?>
 
 <VBox alignment="CENTER" spacing="20.0" xmlns:fx="http://javafx.com/fxml"
-      fx:controller="br.ufrn.imd.circusmanager.Control.TelaListaAnimaisController">
+      fx:controller="br.ufrn.imd.circusmanager.Controller.ZooTelas.TelaListaAnimaisController">
     <padding>
         <Insets bottom="20.0" left="20.0" right="20.0" top="20.0"/>
     </padding>
